Adjust to wxWidgets 3, libupnp-1.6.19, cryptopp562, and Xcode 7

+++ src/ED2KLinkParser.cpp
@@ -185,7 +185,7 @@ string Unescape( const string& str )
 string strip( const string& str )
 {
 	int first = 0;
-	int last  = str.length() - 1;
+	size_t last  = str.length() - 1;
 
 	// A simple but no very optimized way to narrow down the
 	// usable text within the string.
